Organize the code a little
